refactor: 코드 주석 설명 추가 및 달러 원화로 변경 코드 추가(#6)
- 달러 원화 변경 메서드 추가 후 해당 데이터를 사용하는 부분에 모두 추가
- CellDelegate 하나로 합침
- 총 금액 바인딩 시 데이터소스와 다른 컴포넌트와 공유하도록 코드 수정
- 파일 분리 및 폴더링 구조 변경

+++ KioskProject/Presentation/Sources/Model/CartItem.swift
@@ -7,11 +7,12 @@
 
 import Foundation
 
+//MARK: 장바구니 아이템
 struct CartItem: Identifiable, Equatable {
     let product: Product
     var count: Int
-    var totalPrice:Double{
-        return product.price * Double(count)
+    var totalPrice:Int{
+        return product.price * count
     }
     
     var id: String {
